print-servlet - tomcat8 fix
In recent tomcat versions (tested on v8), getRealPath() returns
null if called with a relative argument (which is the case without this
commit).

Adding a leading "/" before calling getRealPath() fixes the NPE
encountered in mapfishapp (georchestra) and geonetwork 3.0.x.

Runtime tested into tomcat8, also tested in jetty to check for any
regression. Both works as expected.